Remove 14 orphaned test fixtures and consolidate temp_checkpoint_dir
- Drop session_tmp_dir, golden_transitions, usaspending_sample from tests/conftest.py — zero consumers found across the entire test suite
- Drop empty_load_metrics, sample_fiscal_awards_df, sample_fiscal_impacts_df, sample_tax_parameters, sample_sbir_investment, sample_discount_rate, sample_fiscal_summary, sample_contracts_df, recent_award, old_award, builder_patents_df from tests/conftest_shared.py — none used as fixture parameters in any test function
- Remove now-orphaned imports: Decimal, timedelta, create_sample_enriched_awards_df
- Remove empty_load_metrics re-export from tests/unit/loaders/conftest.py
- Consolidate identical temp_checkpoint_dir fixture into tests/unit/enrichers/conftest.py; remove local copies from test_chunked_enricher, test_chunked_enrichment, test_chunked_progress
- 4398 unit tests pass, 0 failures
